move queue tests on both `blocking` and `asynk` to separate modules
